Added more item features
Items can now include all custom data, such as names, lore and enchants.
Items can now be added directly from in-game.

+++ FishingLoot/src/main/java/com/ihells/fishingloot/listeners/FishingListener.java
@@ -62,7 +62,7 @@ public class FishingListener implements Listener {
 
         RandomCollection<ItemStack> randomLoot = new RandomCollection<>();
 
-        for (Map.Entry<ItemStack, Double> entrySet : loot.getLoot().entrySet()) {
+        for (Map.Entry<ItemStack, Integer> entrySet : loot.getLoot().entrySet()) {
 
             if (lootUtil.isHighValue(entrySet.getKey())) {
                 if (getLoot(player.getItemInHand()) == 1) {
